elf64 loading, static elf64 support, more ymm / xmm, better flags, and more.

+++ src/emu/hook.rs
@@ -6,10 +6,10 @@ type TypeHookOnInterrupt = fn(emu: &mut emu::Emu, ip_addr: u64, interrupt: u64) 
 // return: allow handle exception?
 type TypeHookOnException = fn(emu: &mut emu::Emu, ip_addr: u64) -> bool;
 // memory read is pre-read you can modify the value that is going to be read.
-type TypeHookOnMemoryRead = fn(emu: &mut emu::Emu, ip_addr: u64, mem_addr: u64, sz: u8);
+type TypeHookOnMemoryRead = fn(emu: &mut emu::Emu, ip_addr: u64, mem_addr: u64, sz: u32);
 // the memory write is pre but you can change the value is going to be written.
 type TypeHookOnMemoryWrite =
-    fn(emu: &mut emu::Emu, ip_addr: u64, mem_addr: u64, sz: u8, value: u128) -> u128;
+    fn(emu: &mut emu::Emu, ip_addr: u64, mem_addr: u64, sz: u32, value: u128) -> u128;
 type TypeHookOnPreInstruction = fn(emu: &mut emu::Emu, ip_addr: u64, ins: &Instruction, sz: usize);
 type TypeHookOnPostInstruction =
     fn(emu: &mut emu::Emu, ip_addr: u64, ins: &Instruction, sz: usize, emu_ok: bool);
